+++ src/main/java/com/commercetools/sync/producttypes/utils/EnumsUpdateActionUtils.java
@@ -10,10 +10,10 @@ import io.sphere.sdk.producttypes.commands.updateactions.RemoveEnumValues;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import java.util.Map;
-import java.util.List;
-import java.util.Optional;
 import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 import java.util.function.BiFunction;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -22,36 +22,11 @@ import static com.commercetools.sync.commons.utils.CommonTypeUpdateActionUtils.b
 import static java.lang.String.format;
 import static java.util.Collections.emptyList;
 import static java.util.Optional.empty;
-import static java.util.Optional.ofNullable;
 import static java.util.Optional.of;
+import static java.util.Optional.ofNullable;
 import static java.util.stream.Collectors.toMap;
 
-public final class ProductTypeUpdateEnumActionsUtils {
-    /**
-     * Given a list of new {@link EnumValue}s, gets a map where the keys are the enum value key, and the values
-     * are the enum instances.
-     *
-     * @param attributeDefinitionName the attribute definition name whose the enum values belong to.
-     * @param enumValues              the list of enum values.
-     * @param <T>                     the enum type of the elements of the list.
-     * @return a map with the enum value key as a key of the map, and the enum
-     *         value as a value of the map.
-     */
-    @Nonnull
-    public static <T extends WithKey> Map<String, T> getEnumValuesKeyMapWithKeyValidation(
-        @Nonnull final String attributeDefinitionName,
-        @Nonnull final List<T> enumValues) {
-
-        return enumValues.stream().collect(
-            toMap(WithKey::getKey, enumValue -> enumValue,
-                (enumValueA, enumValueB) -> {
-                    throw new DuplicateKeyException(format("Enum Values have duplicated keys. "
-                            + "Attribute definition name: '%s', Duplicated enum value: '%s'. "
-                            + "Enum Values are expected to be unique inside their attribute definition.",
-                        attributeDefinitionName, enumValueA.getKey()));
-                }
-            ));
-    }
+final class EnumsUpdateActionUtils {
 
     /**
      * Checks if there are any old enum values which are not existing in the {@code newEnumValues}.
@@ -67,7 +42,7 @@ public final class ProductTypeUpdateEnumActionsUtils {
      *         Otherwise, if the enum values are identical, an empty optional is returned.
      */
     @Nonnull
-    public static <T extends WithKey> Optional<UpdateAction<ProductType>> buildRemoveEnumValuesUpdateActions(
+    static <T extends WithKey> Optional<UpdateAction<ProductType>> buildRemoveEnumValuesUpdateActions(
         @Nonnull final String attributeDefinitionName,
         @Nonnull final List<T> oldEnumValues,
         @Nullable final List<T> newEnumValues) {
@@ -100,7 +75,7 @@ public final class ProductTypeUpdateEnumActionsUtils {
      *         Otherwise, if the enum values order is identical, an empty optional is returned.
      */
     @Nonnull
-    public static <T extends WithKey> Optional<UpdateAction<ProductType>> buildChangeEnumValuesOrderUpdateAction(
+    static <T extends WithKey> Optional<UpdateAction<ProductType>> buildChangeEnumValuesOrderUpdateAction(
         @Nonnull final String attributeDefinitionName,
         @Nonnull final List<T> oldEnumValues,
         @Nonnull final List<T> newEnumValues,
@@ -147,7 +122,7 @@ public final class ProductTypeUpdateEnumActionsUtils {
      *         Otherwise, if the enum values are identical, an empty optional is returned.
      */
     @Nonnull
-    public static <T extends WithKey> List<UpdateAction<ProductType>> buildAddEnumValuesUpdateActions(
+    static <T extends WithKey> List<UpdateAction<ProductType>> buildAddEnumValuesUpdateActions(
         @Nonnull final String attributeDefinitionName,
         @Nonnull final List<T> oldEnumValues,
         @Nonnull final List<T> newEnumValues,
@@ -182,7 +157,7 @@ public final class ProductTypeUpdateEnumActionsUtils {
      *         Otherwise, if the enum values are identical, an empty optional is returned.
      */
     @Nonnull
-    public static <T extends WithKey> List<UpdateAction<ProductType>> buildMatchingEnumValuesUpdateActions(
+    static <T extends WithKey> List<UpdateAction<ProductType>> buildMatchingEnumValuesUpdateActions(
         @Nonnull final String attributeDefinitionName,
         @Nonnull final List<T> oldEnumValues,
         @Nonnull final List<T> newEnumValues,
@@ -205,6 +180,31 @@ public final class ProductTypeUpdateEnumActionsUtils {
             .collect(Collectors.toList());
     }
 
-    private ProductTypeUpdateEnumActionsUtils() {
+    /**
+     * Given a list of new {@link EnumValue}s, gets a map where the keys are the enum value key, and the values
+     * are the enum instances.
+     *
+     * @param attributeDefinitionName the attribute definition name whose the enum values belong to.
+     * @param enumValues              the list of enum values.
+     * @param <T>                     the enum type of the elements of the list.
+     * @return a map with the enum value key as a key of the map, and the enum
+     *         value as a value of the map.
+     */
+    @Nonnull
+    private static <T extends WithKey> Map<String, T> getEnumValuesKeyMapWithKeyValidation(
+            @Nonnull final String attributeDefinitionName,
+            @Nonnull final List<T> enumValues) {
+
+        return enumValues.stream().collect(
+            toMap(WithKey::getKey, enumValue -> enumValue, (enumValueA, enumValueB) -> {
+                    throw new DuplicateKeyException(format("Enum Values have duplicated keys. "
+                                    + "Attribute definition name: '%s', Duplicated enum value: '%s'. "
+                                    + "Enum Values are expected to be unique inside their attribute definition.",
+                            attributeDefinitionName, enumValueA.getKey()));
+                }
+            ));
+    }
+
+    private EnumsUpdateActionUtils() {
     }
 }