+++ src/main/java/com/commercetools/sync/services/impl/ProductTypeServiceImpl.java
@@ -10,11 +10,10 @@ import io.sphere.sdk.producttypes.ProductTypeDraft;
 import io.sphere.sdk.producttypes.commands.ProductTypeCreateCommand;
 import io.sphere.sdk.producttypes.commands.ProductTypeUpdateCommand;
 import io.sphere.sdk.producttypes.queries.ProductTypeQuery;
-import io.sphere.sdk.producttypes.queries.ProductTypeQueryBuilder;
-import io.sphere.sdk.queries.QueryExecutionUtils;
 import org.apache.commons.lang3.StringUtils;
 
 import javax.annotation.Nonnull;
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -23,11 +22,13 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Consumer;
+import java.util.function.Function;
 import java.util.stream.Collectors;
 
 import static java.lang.String.format;
 
 public class ProductTypeServiceImpl implements ProductTypeService {
+    private static final String FETCH_FAILED = "Failed to fetch product types with keys: '%s'. Reason: %s";
     private final BaseSyncOptions syncOptions;
     private final Map<String, String> keyToIdCache = new ConcurrentHashMap<>();
     private boolean isCached = false;
@@ -91,13 +92,31 @@ public class ProductTypeServiceImpl implements ProductTypeService {
 
     @Nonnull
     @Override
-    public CompletionStage<List<ProductType>> fetchMatchingProductsTypesByKeys(@Nonnull final Set<String> keys) {
-        final ProductTypeQuery query = ProductTypeQueryBuilder
-                .of()
-                .plusPredicates(queryModel -> queryModel.key().isIn(keys))
-                .build();
+    public CompletionStage<Set<ProductType>> fetchMatchingProductTypesByKeys(@Nonnull final Set<String> keys) {
+        if (keys.isEmpty()) {
+            return CompletableFuture.completedFuture(Collections.emptySet());
+        }
 
-        return QueryExecutionUtils.queryAll(syncOptions.getCtpClient(), query);
+        final Function<List<ProductType>, List<ProductType>> productTypePageCallBack
+                = productTypePage -> productTypePage;
+
+        return CtpQueryUtils.queryAll(syncOptions.getCtpClient(),
+                ProductTypeQuery.of().withPredicates(queryModel -> queryModel.key().isIn(keys)),
+                productTypePageCallBack)
+                            .handle((fetchedProductTypes, sphereException) -> {
+                                if (sphereException != null) {
+                                    syncOptions
+                                            .applyErrorCallback(format(FETCH_FAILED, keys, sphereException),
+                                                    sphereException);
+                                    return Collections.emptySet();
+                                }
+                                return fetchedProductTypes.stream()
+                                                          .flatMap(List::stream)
+                                                          .peek(productType ->
+                                                                  keyToIdCache.put(productType.getKey(),
+                                                                          productType.getId()))
+                                                          .collect(Collectors.toSet());
+                            });
     }
 
     @Nonnull