+++ src/test/java/com/commercetools/sync/producttypes/utils/producttypeactionutils/BuildAttributeDefinitionUpdateActionsTest.java
@@ -58,7 +58,8 @@ public class BuildAttributeDefinitionUpdateActionsTest {
         RES_ROOT + "product-type-with-attribute-definitions-cbd.json";
     private static final String PRODUCT_TYPE_WITH_ATTRIBUTES_ABC_WITH_DIFFERENT_TYPE =
         RES_ROOT + "product-type-with-attribute-definitions-abc-with-different-type.json";
-
+    private static final String PRODUCT_TYPE_WITH_ATTRIBUTES_ABC_WITHOUT_ATTRIBUTE_TYPE =
+            RES_ROOT + "product-type-with-attribute-definitions-abc-without-attribute-type.json";
 
     private static final ProductTypeSyncOptions SYNC_OPTIONS = ProductTypeSyncOptionsBuilder
         .of(mock(SphereClient.class))
@@ -442,5 +443,75 @@ public class BuildAttributeDefinitionUpdateActionsTest {
                 .isSearchable(true)
                 .build())
         );
+    }
+
+    @Test
+    public void buildAttributesUpdateActions_WithoutOldAttributeType_ShouldNotBuildActionsAndTriggerErrorCallback() {
+        final ProductType oldProductType =
+                readObjectFromResource(PRODUCT_TYPE_WITH_ATTRIBUTES_ABC_WITHOUT_ATTRIBUTE_TYPE, ProductType.class);
+
+        final ProductTypeDraft newProductTypeDraft = readObjectFromResource(
+                PRODUCT_TYPE_WITH_ATTRIBUTES_ABC,
+                ProductTypeDraft.class
+        );
+
+        final List<String> errorMessages = new ArrayList<>();
+        final List<Throwable> exceptions = new ArrayList<>();
+        final ProductTypeSyncOptions syncOptions =
+                ProductTypeSyncOptionsBuilder.of(mock(SphereClient.class))
+                                             .errorCallback((errorMessage, exception) -> {
+                                                 errorMessages.add(errorMessage);
+                                                 exceptions.add(exception);
+                                             })
+                                             .build();
+
+        final List<UpdateAction<ProductType>> updateActions = buildAttributesUpdateActions(
+                oldProductType,
+                newProductTypeDraft,
+                syncOptions
+        );
+
+        assertThat(updateActions).isEmpty();
+        assertThat(errorMessages).hasSize(1);
+        assertThat(errorMessages.get(0)).matches("Failed to build update actions for the attributes definitions of the "
+                + "product type with the key 'key'. Reason: .*BuildUpdateActionException: "
+                + "Attribute type is not set for the old attribute definition.");
+        assertThat(exceptions).hasSize(1);
+        assertThat(exceptions.get(0)).isExactlyInstanceOf(BuildUpdateActionException.class);
+    }
+
+    @Test
+    public void buildAttributesUpdateActions_WithoutAttributeType_ShouldNotBuildActionsAndTriggerErrorCallback() {
+        final ProductType oldProductType =
+                readObjectFromResource(PRODUCT_TYPE_WITH_ATTRIBUTES_ABC, ProductType.class);
+
+        final ProductTypeDraft newProductTypeDraft = readObjectFromResource(
+                PRODUCT_TYPE_WITH_ATTRIBUTES_ABC_WITHOUT_ATTRIBUTE_TYPE,
+                ProductTypeDraft.class
+        );
+
+        final List<String> errorMessages = new ArrayList<>();
+        final List<Throwable> exceptions = new ArrayList<>();
+        final ProductTypeSyncOptions syncOptions =
+                ProductTypeSyncOptionsBuilder.of(mock(SphereClient.class))
+                                             .errorCallback((errorMessage, exception) -> {
+                                                 errorMessages.add(errorMessage);
+                                                 exceptions.add(exception);
+                                             })
+                                             .build();
+
+        final List<UpdateAction<ProductType>> updateActions = buildAttributesUpdateActions(
+                oldProductType,
+                newProductTypeDraft,
+                syncOptions
+        );
+
+        assertThat(updateActions).isEmpty();
+        assertThat(errorMessages).hasSize(1);
+        assertThat(errorMessages.get(0)).matches("Failed to build update actions for the attributes definitions of the "
+                + "product type with the key 'key'. Reason: .*BuildUpdateActionException: "
+                + "Attribute type is not set for the new/draft attribute definition.");
+        assertThat(exceptions).hasSize(1);
+        assertThat(exceptions.get(0)).isExactlyInstanceOf(BuildUpdateActionException.class);
     }
 }